Wait for each example to finish before continuing.
Fixes #3.

With Concordion 2.0, each example is run as a separate JUnit test.
Run commands inside an example are launched as tasks on separate threads when using the parallel runner.
In order for the JUnit results to be accurate, we must wait at the end of each example for all of its run commands to finish.
This includes the "outer" example, which contains all commands in the specification that are not in a specific example command.

Where a specification contains multiple examples, each example becomes a synchronisation point,
and further examples will not be executed until the current example is finished.

Examples within a specification are run serially.
Parallel running of examples in a specification is not supported since it would not be thread-safe to do so.
For example, the XOM library is not thread-safe and is used by many commands within each example.